修改头像保存后不显示的bug

+++ src/main/java/cn/zzuzl/xblog/util/FileUtil.java
@@ -13,7 +13,7 @@ public class FileUtil {
     private static Logger logger = LogManager.getLogger(FileUtil.class);
 
     // 清理目录文件
-    public static int cleanFiles(String folder) {
+    public static int cleanFiles(String folder, long timeGap) {
         int count = 0;
         if (StringUtils.isEmpty(folder)) {
             return count;
@@ -21,16 +21,18 @@ public class FileUtil {
         File rootFolder = new File(folder);
 
         if (rootFolder.exists()) {
-            // 不是目录则删除
             if (!rootFolder.isDirectory()) {
-                rootFolder.deleteOnExit();
-                logger.info("清理:" + rootFolder.getAbsolutePath());
-                count ++;
+                // 删除30分钟之前的文件
+                if (System.currentTimeMillis() - rootFolder.lastModified() > timeGap) {
+                    rootFolder.deleteOnExit();
+                    logger.info("清理:" + rootFolder.getAbsolutePath());
+                    count++;
+                }
             } else {
                 File[] files = rootFolder.listFiles();
                 if (files != null && files.length > 0) {
                     for (File file : files) {
-                        count += cleanFiles(file.getAbsolutePath());
+                        count += cleanFiles(file.getAbsolutePath(), timeGap);
                     }
                 }
             }
